refactoring using blueprints
