+++ core/users/models.py
@@ -3,7 +3,7 @@
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from application import db
+from core import db
 
 
 class User(db.Model, UserMixin):
