+++ requirements.in
@@ -1,6 +1,7 @@
 email-validator~=2.2.0
 flask-login~=0.6.3
 flask-sqlalchemy~=3.1.1
+psycopg2~=2.9.10
 python-slugify~=8.0.4
 Flask~=3.1.0
 Flask-WTF~=1.2.2
